requirements.txt
requests
pandas
numpy
pyTelegramBotAPI
python-binance
tensorflow
scikit-learn
python-dotenv

